\name{SynIntTest-package}
\alias{SynIntTest-package}
\alias{SynIntTest}
\docType{package}
\title{
\packageTitle{SynIntTest}
}
\description{
\packageDescription{SynIntTest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SynIntTest}
\packageIndices{SynIntTest}
SIMsampler samples from the posterior distribution of the synergistic interaction model.
}
\author{
\packageAuthor{SynIntTest}

Maintainer: \packageMaintainer{SynIntTest}
}
\references{
[to be added]
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
\code{\link{SIMsampler}}
}
\examples{
#### To be added ####
}
