\name{SynIntTestExact-package}
\alias{SynIntTestExact-package}
\alias{SynIntTestExact}
\docType{package}
\title{
\packageTitle{SynIntTestExact}
}
\description{
\packageDescription{SynIntTestExact}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SynIntTestExact}
\packageIndices{SynIntTestExact}
SIMsampler samples from the posterior distribution of the synergistic interaction model.
}
\author{
\packageAuthor{SynIntTestExact}

Maintainer: \packageMaintainer{SynIntTestExact}
}
\references{
[to be added]
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
\code{\link{SIMsampler}}
}
\examples{
#### To be added ####
}
